[MOD-2069] Investigate xctool for building application using Travis scripts

+++ travis/script.sh
@@ -27,34 +27,46 @@ if [ -d "$MODULE_ROOT/ios/" ]; then
 
 		#Install module
 		echo "Installing module"
-		ti sdk install $MODULE_ID*
-
-		#This already happens in the previous script
-		#echo "Downloading latest SDK"
-		#ti sdk install
-		#ti sdk select latest
+		ti sdk install $MODULE_ID* -q
 
 		#Create new application
 		echo "Creating new application"
-		ti create -t app -p ios -d "./TestModule" -n "TestModule" --id "com.appc.TestModule" -u "http://appcelerator.com" --force
+		ti create -t app -p ios -d "./TestModule" -n "TestModule" --id "com.appc.TestModule" -u "http://appcelerator.com" --force -q
 		cd ./TestModule/TestModule
 
 		#Append module to manifest
-		echo "Add module to application"
+		echo "Adding module to application"
 		sed -i "" 's/<modules>/&<module version="'$MODULE_VERSION'">'$MODULE_ID'<\/module>/g' tiapp.xml
 
-		#echo "Copying module example"
-		#cp -r $HOME/Library/Application\ Support/Titanium/modules/iphone/$MODULE_ID/$MODULE_VERSION/example/. ./Resources/
-
-		#Append module to app.js
-		cd ./Resources/
-		echo -e "var moduleToTest = require('"$MODULE_ID"');\n$(cat app.js)" > app.js
-		cd ../
+		#Configure app.js
+		MODULE_EXAMPLE=$HOME/Library/Application\ Support/Titanium/modules/iphone/$MODULE_ID/$MODULE_VERSION/example/
+		if [ -e "$MODULE_EXAMPLE/app.js" ]; then
+			echo "Copying module example"
+			cp -r "$MODULE_EXAMPLE." ./Resources/
+		else
+			#Append module to app.js
+			echo "Append module require to app.js"
+			cd ./Resources/
+			echo -e "var moduleToTest = require('"$MODULE_ID"');\n$(cat app.js)" > app.js
+			cd ../
+		fi
 
 		#Build application but do not run simulator
-		echo "Build application"
-		ti build -b -p ios -d "./"
-	
+		echo "Building application"
+		ti build -b -p ios -d "./" --no-banner --no-progress-bars --no-prompt -q
+		
+		#xctool
+		cd ./build/iphone
+		echo "Building scheme"
+		#need to build scheme
+		#need to find a better way to build scheme
+		open TestModule.xcodeproj
+		sleep 5
+		osascript -e "tell application \"Xcode\" to quit"
+		#xcodebuild -list
+		echo "Testing application using xctool"
+		xctool -scheme TestModule build-tests
+
 		let STATUS=$?
 		
 		if (( "$RETSTATUS" == "0" )) && (( "$STATUS" != "0" )); then
